+++ index.js
@@ -1,173 +1,12 @@
-require('dotenv').config() //enviornment variables (something has to be installed before we can use this)
-const express = require('express')
-const cors = require('cors')
-const app = express()
-const Note = require('./models/note') //mongo db code
+//AFTER RESTRUCTURING index2.js
 
-//MIDDLEWARES
-app.use(express.static('build')) //this has to be declared when deploying the server to internet. so that requests goes to 'build' dir first
-app.use(cors()) //cors policy
-app.use(express.json())//parse json 
+const app = require('./app') // the actual Express application
+const http = require('http')
+const config = require('./utils/config')
+const logger = require('./utils/logger')
 
-// let notes = [
-//     {
-//       id: 1,
-//       content: "HTML is easy",
-//       date: "2019-05-30T17:30:31.098Z",
-//       important: true
-//     },
-//     {
-//       id: 2,
-//       content: "Browser can execute only Javascript",
-//       date: "2019-05-30T18:39:34.091Z",
-//       important: false
-//     },
-//     {
-//       id: 3,
-//       content: "GET and POST are the most important methods of HTTP protocol",
-//       date: "2019-05-30T19:20:14.298Z",
-//       important: true
-//     }
-//   ]
+const server = http.createServer(app)
 
-
-
-app.get('/', (request, response) => {
-  response.send('<h1>Hello World!</h1>')
-})
-
-app.get('/api/notes', (request, response) => {
-  Note.find({}).then(notes => {
-    response.json(notes)
-  })
-})
-
-//BEFORE MONGODB
-// app.get('/api/notes/:id', (request, response) => {
-//     const id = Number(request.params.id)
-//     const note = notes.find(note => note.id === id)
-    
-//     if (note) {
-//         response.json(note)
-//       } else {
-//         response.status(404).end()
-//       }
-//     })
-
-app.get('/api/notes/:id', (request, response, next) => {
-  Note.findById(request.params.id)
-    .then(note => {
-      if (note) {
-        response.json(note)
-      } else {
-        response.status(404).end()
-      }
-    })
-    .catch(error => next(error))
-})
-
-// app.delete('/api/notes/:id', (request, response) => {
-//   const id = Number(request.params.id)
-//   notes = notes.filter(note => note.id !== id)
-
-//   response.status(204).end()
-// })
-
-app.delete('/api/notes/:id', (request, response, next) => {
-  Note.findByIdAndRemove(request.params.id)
-    .then(result => {
-      response.status(204).end()
-    })
-    .catch(error => next(error))
-})
-
-
-const generateId = () => {
-  const maxId = notes.length > 0
-    ? Math.max(...notes.map(n => n.id))
-    : 0
-  return maxId + 1
-}
-
-//BEFORE MONGODB
-// app.post('/api/notes', (request, response) => {
-//   const body = request.body
-
-//   if (!body.content) {
-//     return response.status(400).json({ 
-//       error: 'content missing' 
-//     })
-//   }
-
-//   const note = {
-//     content: body.content,
-//     important: body.important || false,
-//     date: new Date(),
-//     id: generateId(),
-//   }
-
-//   notes = notes.concat(note)
-
-//   response.json(note)
-// })
-
-app.post('/api/notes', (request, response) => {
-  const body = request.body
-
-  if (body.content === undefined) {
-    return response.status(400).json({ error: 'content missing' })
-  }
-
-  const note = new Note({
-    content: body.content,
-    important: body.important || false,
-    date: new Date(),
-  })
-
-  note.save().then(savedNote => {
-    response.json(savedNote)
-  })
-})
-
-app.put('/api/notes/:id', (request, response, next) => {
-  const body = request.body
-
-  const note = {
-    content: body.content,
-    important: body.important,
-  }
-
-  //By default, the updatedNote parameter of the event handler receives the original document without the modifications. We added the optional { new: true }parameter, which will cause our event handler to be called with the new modified document instead of the original.
-  Note.findByIdAndUpdate(request.params.id, note, { new: true })
-    .then(updatedNote => {
-      response.json(updatedNote)
-    })
-    .catch(error => next(error))
-})
-
-//ERROR HANDLING MIDDLEWARES SHOULD BE DEFINED LAST
-const unknownEndpoint = (request, response) => {
-  response.status(404).send({ error: 'unknown endpoint' })
-}
-
-// handler of requests with unknown endpoint
-app.use(unknownEndpoint)
-
-const errorHandler = (error, request, response, next) => {
-  console.error(error.message)
-
-  if (error.name === 'CastError') {
-    console.log("a malformtted request came")
-    return response.status(400).send({ error: 'malformatted id' })
-  } 
-  //passes the error forward to the default Express error handler.
-  next(error)
-}
-
-// handler of requests with result to errors
-app.use(errorHandler)
-
-const PORT = process.env.PORT || 3001
-app.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`)
+server.listen(config.PORT, () => {
+  logger.info(`Server running on port ${config.PORT}`)
 })
