+++ models/note.js
@@ -1,26 +1,42 @@
-require('dotenv').config()
+//defining mongodb scheme, connection code is move to app.js
 
 const mongoose = require('mongoose')
 
-const url = process.env.MONGODB_URI
+// const url = process.env.MONGODB_URI
 
-console.log('connecting to', url)
+// console.log('connecting to', url)
 
-mongoose.connect(url)
-  .then(result => {
-    console.log('connected to MongoDB')
-  })
-  .catch((error) => {
-    console.log('error connecting to MongoDB:', error.message)
-  })
+// mongoose.connect(url)
+//   .then(result => {
+//     console.log('connected to MongoDB')
+//   })
+//   .catch((error) => {
+//     console.log('error connecting to MongoDB:', error.message)
+//   })
 
-  //transform the objects returnd from db into string and delete unwanted fields
+//schema 
+// const noteSchema = new mongoose.Schema({
+//   content: String,
+//   date: Date,
+//   important: Boolean,
+// })
+
+//schema with validation
 const noteSchema = new mongoose.Schema({
-  content: String,
-  date: Date,
+  content: {
+    type: String,
+    required: true,
+    minlength: 5
+  },
+  date: {
+    type: Date,
+    required: true,
+  },
   important: Boolean,
 })
 
+
+//transform the objects returnd from db into string and delete unwanted fields
 noteSchema.set('toJSON', {
   transform: (document, returnedObject) => {
     returnedObject.id = returnedObject._id.toString()
